// src/types/api/service/models/InquiryDto.ts
export interface InquiryDto {
    /**
     * 
     * @type {number}
     * @memberof InquiryDto
     */
    inquiryId?: number;
    /**
     * 
     * @type {string}
     * @memberof InquiryDto
     */
    whyCode?: string;
    /**
     * 
     * @type {string}
     * @memberof InquiryDto
     */
    title?: string;
    /**
     * 
     * @type {string}
     * @memberof InquiryDto
     */
    content?: string;
    /**
     * 
     * @type {string}
     * @memberof InquiryDto
     */
    stateCode?: string;
    /**
     * 
     * @type {string}
     * @memberof InquiryDto
     */
    email?: string;
    /**
     * 
     * @type {string}
     * @memberof InquiryDto
     */
    mobile?: string;
    /**
     * 
     * @type {string}
     * @memberof InquiryDto
     */
    phone?: string;
    /**
     * 
     * @type {number}
     * @memberof InquiryDto
     */
    companyId?: number;
    /**
     * 
     * @type {number}
     * @memberof InquiryDto
     */
    licenseId?: number;
    /**
     * 
     * @type {number}
     * @memberof InquiryDto
     */
    userId?: number;
    /**
     * 
     * @type {number}
     * @memberof InquiryDto
     */
    journalistId?: number;
    /**
     * 
     * @type {number}
     * @memberof InquiryDto
     */
    mediaId?: number;
    /**
     * 
     * @type {number}
     * @memberof InquiryDto
     */
    newsId?: number;
    /**
     * 
     * @type {string}
     * @memberof InquiryDto
     */
    targetTitle?: string;
    /**
     * 
     * @type {Array<AttachedDto>}
     * @memberof InquiryDto
     */
    attachedList?: Array<AttachedDto>;
    /**
     * 
     * @type {Array<InquiryResponseDto>}
     * @memberof InquiryDto
     */
    inquiryResponseList?: Array<InquiryResponseDto>;
    /**
     * 
     * @type {string}
     * @memberof InquiryDto
     */
    regisAt?: string;
    /**
     * 
     * @type {string}
     * @memberof InquiryDto
     */
    updateAt?: string;
    /**
     * 
     * @type {string}
     * @memberof InquiryDto
     */
    completeAt?: string;
}