// src/types/api/service/models/PageInquiryDtoForList.ts
export interface PageInquiryDtoForList {
    /**
     * 
     * @type {number}
     * @memberof PageInquiryDtoForList
     */
    totalElements?: number;
    /**
     * 
     * @type {number}
     * @memberof PageInquiryDtoForList
     */
    totalPages?: number;
    /**
     * 
     * @type {boolean}
     * @memberof PageInquiryDtoForList
     */
    first?: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof PageInquiryDtoForList
     */
    last?: boolean;
    /**
     * 
     * @type {number}
     * @memberof PageInquiryDtoForList
     */
    size?: number;
    /**
     * 
     * @type {Array<InquiryDtoForList>}
     * @memberof PageInquiryDtoForList
     */
    content?: Array<InquiryDtoForList>;
    /**
     * 
     * @type {number}
     * @memberof PageInquiryDtoForList
     */
    number?: number;
    /**
     * 
     * @type {SortObject}
     * @memberof PageInquiryDtoForList
     */
    sort?: SortObject;
    /**
     * 
     * @type {number}
     * @memberof PageInquiryDtoForList
     */
    numberOfElements?: number;
    /**
     * 
     * @type {PageableObject}
     * @memberof PageInquiryDtoForList
     */
    pageable?: PageableObject;
    /**
     * 
     * @type {boolean}
     * @memberof PageInquiryDtoForList
     */
    empty?: boolean;
}